t7

+++ projects/ngx-dynamic-pages/postinstall.js
@@ -19,28 +19,32 @@ const preJson = `[
 ]`;
 
 // Caminho do arquivo do projeto que está instalando a biblioteca
-const caminhoArquivoProjeto = path.resolve(process.cwd(), 'src/app/app.module.ts');
+const caminhoArquivoProjeto = path.resolve(process.cwd(), '../../src/app/app.module.ts');
 const importacaoBiblioteca = `import { NgxDynamicPagesModule } from 'ngx-dynamic-pages';`;
 
-// Navega para fora do diretório node_modules
-process.chdir('../../');
+// Verifica se o arquivo existe antes de tentar manipulá-lo
+if (fs.existsSync(caminhoArquivoProjeto)) {
+  // Lê o código do projeto
+  const codigoProjeto = fs.readFileSync(caminhoArquivoProjeto, 'utf-8');
 
-// Lê o código do projeto
-const codigoProjeto = fs.readFileSync(caminhoArquivoProjeto, 'utf-8');
+  // Procura por uma linha que começa com "import"
+  const importMatch = codigoProjeto.match(/import.*;/);
 
-// Procura por uma linha que começa com "import"
-const importMatch = codigoProjeto.match(/import.*;/);
+  // Adiciona a importação logo após a última importação encontrada
+  const novoCodigoProjeto = importMatch
+    ? codigoProjeto.replace(importMatch[0], `${importMatch[0]}\n${importacaoBiblioteca}`)
+    : codigoProjeto + `\n${importacaoBiblioteca}`;
 
-// Adiciona a importação logo após a última importação encontrada
-const novoCodigoProjeto = importMatch
-  ? codigoProjeto.replace(importMatch[0], `${importMatch[0]}\n${importacaoBiblioteca}`)
-  : codigoProjeto + `\n${importacaoBiblioteca}`;
+  // Escreve o novo código no arquivo do projeto
+  fs.writeFileSync(caminhoArquivoProjeto, novoCodigoProjeto, 'utf-8');
 
-// Escreve o novo código no arquivo do projeto
-fs.writeFileSync(caminhoArquivoProjeto, novoCodigoProjeto, 'utf-8');
+  console.log(`Arquivo ${caminhoArquivoProjeto} atualizado com sucesso!`);
+} else {
+  console.error(`Erro: Arquivo ${caminhoArquivoProjeto} não encontrado.`);
+}
 
 // Caminho do arquivo a ser gerado
-const caminhoArquivo = path.resolve(process.cwd(), 'src/jsonParameterize.json');
+const caminhoArquivo = path.resolve(process.cwd(), '../../src/jsonParameterize.json');
 
 // Escreve as informações no arquivo jsonParameterize.json
 fs.writeFileSync(caminhoArquivo, preJson, 'utf-8');
